fix: 충돌 오류 해결 (#301)
