+++ src/main/java/com/soda/project/application/ProjectFacade.java
@@ -3,10 +3,11 @@ package com.soda.project.application;
 import com.querydsl.core.Tuple;
 import com.soda.global.log.data.annotation.LoggableEntityAction;
 import com.soda.global.response.GeneralException;
-import com.soda.member.entity.Company;
-import com.soda.member.entity.Member;
-import com.soda.member.service.CompanyService;
-import com.soda.member.service.MemberService;
+import com.soda.member.application.CompanyFacade;
+import com.soda.member.domain.Member;
+import com.soda.member.domain.MemberService;
+import com.soda.member.domain.company.Company;
+import com.soda.member.domain.company.CompanyService;
 import com.soda.project.application.validator.ProjectValidator;
 import com.soda.project.domain.Project;
 import com.soda.project.domain.ProjectErrorCode;
