+++ src/main/java/com/soda/project/application/validator/ProjectValidator.java
@@ -3,6 +3,8 @@ package com.soda.project.application.validator;
 import com.soda.global.response.CommonErrorCode;
 import com.soda.global.response.GeneralException;
 import com.soda.member.domain.*;
+import com.soda.member.domain.company.Company;
+import com.soda.member.domain.company.CompanyService;
 import com.soda.project.domain.Project;
 import com.soda.project.domain.ProjectErrorCode;
 import com.soda.project.domain.member.MemberProjectRole;
