+++ src/main/java/com/soda/project/domain/ProjectService.java
@@ -2,8 +2,8 @@ package com.soda.project.domain;
 
 import com.querydsl.core.Tuple;
 import com.soda.global.response.GeneralException;
-import com.soda.member.domain.Company;
 import com.soda.member.domain.Member;
+import com.soda.member.domain.company.Company;
 import com.soda.project.domain.company.CompanyProjectRole;
 import com.soda.project.interfaces.dto.ProjectListResponse;
 import com.soda.project.interfaces.dto.ProjectSearchCondition;
