+++ src/main/java/com/soda/project/domain/member/MemberProjectProvider.java
@@ -1,7 +1,7 @@
 package com.soda.project.domain.member;
 
-import com.soda.member.domain.Company;
 import com.soda.member.domain.Member;
+import com.soda.member.domain.company.Company;
 import com.soda.project.domain.Project;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
