+++ src/main/java/com/soda/project/infrastructure/MemberProjectProviderImpl.java
@@ -1,7 +1,7 @@
 package com.soda.project.infrastructure;
 
-import com.soda.member.domain.Company;
 import com.soda.member.domain.Member;
+import com.soda.member.domain.company.Company;
 import com.soda.project.domain.Project;
 import com.soda.project.domain.member.MemberProject;
 import com.soda.project.domain.member.MemberProjectProvider;
