+++ src/main/java/com/soda/project/interfaces/dto/DevCompanyAssignmentResponse.java
@@ -1,8 +1,5 @@
 package com.soda.project.interfaces.dto;
 
-import com.soda.member.domain.company.Company;
-import com.soda.member.domain.Member;
-import com.soda.project.domain.member.enums.MemberProjectRole;
 import lombok.Builder;
 import lombok.Getter;
