+++ src/main/java/com/soda/project/interfaces/dto/company/CompanyProjectDTO.java
@@ -1,8 +1,6 @@
 package com.soda.project.interfaces.dto.company;
 
 import com.soda.member.domain.company.Company;
-import com.soda.project.domain.company.enums.CompanyProjectRole;
-import com.soda.member.domain.Company;
 import com.soda.project.domain.company.CompanyProjectRole;
 import com.soda.project.domain.company.CompanyProject;
 import com.soda.project.domain.Project;
